pyproject.toml
[project]
name = "gruporom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-crispy-forms>=2.4",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.3",
    "fabric>=3.2.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.test_settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = [
    "--reuse-db",
    "--ds=core.test_settings",
    "-v",
    "--tb=short",
    "--strict-markers",
]
testpaths = ["core/tests"]
